Pool exhaustion safety, stage overflow, FPS divider overflow
- AnimClass/BulletClass operator new marked noexcept (returns NULL on pool full instead of UB)
- Stage counter widened from unsigned short to unsigned int (prevents wraparound after ~36 min at 30 FPS)
- FPSDivider periodic reset prevents accumulator overflow in debug display

+++ modern.src/stage.h
@@ -44,7 +44,7 @@ class StageClass {
 		**	This handles the animation stage of the object. This includes smoke, walking,
 		**	flapping, and rocket flames.
 		*/
-		unsigned short Stage;
+		unsigned int Stage;
 
 		/*
 		**	This is the countdown timer for stage animation. When this counts down
